perf: 解耦插件管理页面
