+++ packages/web/src/components/plugin/admin/index.tsx
@@ -1,0 +1,23 @@
+import PluginRow, { type PluginItem, type PluginRowProps } from './PluginRow'
+import FilterCards, { type PluginCounts, type FilterCardsProps } from './FilterCards'
+import TableContent, { type TableContentProps } from './TableContent'
+import { getUpdateStatusConfig, getTypeConfig, renderIcon, stopPropagation, type PluginType } from './utils'
+
+export {
+  PluginRow,
+  FilterCards,
+  TableContent,
+  getUpdateStatusConfig,
+  getTypeConfig,
+  renderIcon,
+  stopPropagation,
+}
+
+export type {
+  PluginItem,
+  PluginRowProps,
+  PluginCounts,
+  FilterCardsProps,
+  TableContentProps,
+  PluginType,
+}
