+++ packages/web/src/pages/dashboard/plugins/manage.tsx
@@ -1,374 +1,12 @@
 /* eslint-disable @stylistic/indent */
-import { useState, useCallback, ReactElement, useEffect, memo, useRef, useMemo } from 'react'
+import { useState, useCallback, ReactElement, useEffect, useMemo } from 'react'
 import { useSearchParams } from 'react-router-dom'
 import { useRequest } from 'ahooks'
-import { Card } from '@heroui/card'
-import { Chip } from '@heroui/chip'
 import { Button } from '@heroui/button'
-import { Spinner } from '@heroui/spinner'
-import { Checkbox } from '@heroui/checkbox'
-import { Progress } from '@heroui/progress'
-import { Dropdown, DropdownTrigger, DropdownMenu, DropdownItem } from '@heroui/dropdown'
-import { TbApps, TbBrandGit, TbCircleCheck, TbCircleDashed, TbArrowUp, TbRefresh, TbTrash, TbArrowsUp } from 'react-icons/tb'
-import { IoSettingsOutline } from 'react-icons/io5'
+import { TbCircleCheck, TbCircleDashed, TbArrowsUp, TbRefresh, TbTrash } from 'react-icons/tb'
 import { MdOutlineExtension } from 'react-icons/md'
-import { fetchPluginList, type PluginItem, type PluginType } from '@/mock/pluginManageData'
-import StatCard from '@/components/card/StatCard'
-import LazyPluginLoader from '@/components/plugins/LazyPluginLoader'
-
-/**
- * 获取更新状态配置
- * @param status - 更新状态
- * @param type - 插件类型
- * @returns 状态配置
- */
-const getUpdateStatusConfig = (type: PluginType, version: string, latesHash: string) => {
-  if (type === 'app') {
-    return {
-      text: '-',
-      color: 'default' as const,
-      icon: <TbCircleDashed className='text-xs mr-1' />,
-    }
-  }
-
-  if (version === latesHash) {
-    return {
-      text: '最新',
-      color: 'success' as const,
-      icon: <TbCircleCheck className='text-xs mr-1' />,
-    }
-  }
-
-  if (version !== latesHash) {
-    return {
-      text: '可更新',
-      color: 'warning' as const,
-      icon: <TbArrowUp className='text-xs mr-1' />,
-    }
-  }
-
-  return {
-    text: '未知',
-    color: 'default' as const,
-    icon: <TbCircleDashed className='text-xs mr-1' />,
-  }
-}
-
-/**
- * 获取类型配置
- * @param type - 插件类型
- * @returns 类型配置信息
- */
-const getTypeConfig = (type: PluginType) => {
-  switch (type) {
-    case 'git':
-      return {
-        color: 'warning',
-        text: 'Git插件',
-      }
-    case 'app':
-      return {
-        color: 'primary',
-        text: 'App插件',
-      }
-    case 'all':
-    default:
-      return {
-        color: 'default',
-        text: '全部',
-      }
-  }
-}
-
-/**
- * 渲染图标的帮助函数
- * @param size - 图标尺寸
- * @param Icon - 图标组件
- * @returns 渲染的图标组件
- */
-const renderIcon = (size: number, Icon: React.ComponentType<{ size?: number }>) => {
-  return <Icon size={size} />
-}
-
-/**
- * 阻止事件冒泡
- * @param e - 事件对象
- */
-const stopPropagation = (e: React.MouseEvent | React.TouchEvent) => {
-  e.stopPropagation()
-}
-
-/**
- * 插件行组件，使用memo优化渲染
- */
-const PluginRow = memo(({
-  plugin,
-  isSelected,
-  onSelect,
-  openSettings,
-}: {
-  plugin: PluginItem
-  isSelected: boolean
-  onSelect: (id: string, isSelected: boolean) => void
-  openSettings: (pluginId: string) => void
-}) => {
-  /** 插件更新状态 */
-  const statusConfig = getUpdateStatusConfig(plugin.type, plugin.version, plugin.latestHash)
-  /** 插件类型配置 */
-  const typeConfig = getTypeConfig(plugin.type)
-
-  /** 处理选择状态变更 */
-  const handleSelectionChange = useCallback(() => {
-    onSelect(plugin.id, !isSelected)
-  }, [plugin.id, isSelected, onSelect])
-
-  /** 处理下拉菜单操作 */
-  const handleDropdownAction = useCallback((key: React.Key) => {
-    console.log(`对插件 ${plugin.id} 执行操作: ${key}`)
-    // 根据key类型处理不同的操作
-    switch (key) {
-      case 'settings':
-        openSettings(plugin.id)
-        break
-      case 'update':
-        // TODO:  处理更新操作
-        break
-      case 'forceUpdate':
-        // TODO: 处理强制更新操作
-        break
-      case 'uninstall':
-        // TODO: 处理卸载操作
-        break
-      default:
-        break
-    }
-  }, [plugin.id, openSettings])
-
-  return (
-    <div className='grid grid-cols-12 w-full border-b border-default-100/40 hover:bg-default-50/70 transition-colors' onClick={() => onSelect(plugin.id, !isSelected)}>
-      {/* 选择框 */}
-      <div className='py-3 md:py-4 px-2 md:px-4 text-sm flex items-center justify-center col-span-1' onClick={stopPropagation}>
-        <div className='flex items-center justify-center w-full'>
-          <Checkbox
-            isSelected={isSelected}
-            onValueChange={handleSelectionChange}
-            size='sm'
-            aria-label={`选择 ${plugin.name}`}
-            classNames={{
-              base: 'w-4 h-4',
-              wrapper: 'rounded-full w-4 h-4 border-1 border-default-300 data-[selected=true]:border-blue-500 data-[selected=true]:bg-blue-500 data-[hover=true]:border-blue-400 data-[hover=true]:bg-blue-400/20 transition-all',
-              icon: 'text-white text-[10px]',
-            }}
-          />
-        </div>
-      </div>
-
-      {/* 插件名称 */}
-      <div className='py-3 md:py-4 px-2 md:px-4 text-sm col-span-6 sm:col-span-4 flex items-center'>
-        <div className='flex items-center w-full'>
-          <div className='flex-shrink-0'>
-            <div
-              className='w-2 md:w-2.5 h-2 md:h-2.5 rounded-full mr-1.5 md:mr-2.5 flex-shrink-0 translate-y-[0.5px] sm:translate-y-0'
-              style={{
-                backgroundColor: plugin.type === 'app' ? '#10b981' : '#f59e0b',
-                boxShadow: plugin.type === 'app' ? '0 0 6px rgba(16, 185, 129, 0.3)' : '0 0 6px rgba(245, 158, 11, 0.3)',
-              }}
-            />
-          </div>
-          <div className='font-medium text-default-700 dark:text-default-300 w-full break-words whitespace-normal'>
-            {window.innerWidth <= 640
-              ? (
-                <span className='text-xs'>
-                  {
-                    plugin.type === 'app'
-                      ? plugin.name.replace('karin-plugin-', '')
-                      : plugin.name
-                  }
-                </span>
-              )
-              : plugin.name}
-          </div>
-        </div>
-      </div>
-
-      {/* 当前版本 */}
-      <div className='hidden sm:block py-3 md:py-4 px-1 sm:px-2 md:px-4 text-sm col-span-3 sm:col-span-2 text-center'>
-        {plugin.type === 'app'
-          ? (
-            <span className='text-xs text-default-500'>-</span>
-          )
-          : (
-            <Chip size='sm' variant='flat' color='primary' className='bg-primary-100/50'>
-              {plugin.version}
-            </Chip>
-          )}
-      </div>
-
-      {/* 最新版本 */}
-      <div className='hidden sm:block py-3 md:py-4 px-1 sm:px-2 md:px-4 text-sm col-span-3 sm:col-span-2 text-center'>
-        {plugin.type === 'app'
-          ? (
-            <span className='text-xs text-default-500'>-</span>
-          )
-          : (
-            plugin.latestHash !== plugin.version
-              ? (
-                <Chip size='sm' variant='flat' color='warning' className='bg-warning-100/50'>
-                  {plugin.latestHash}
-                </Chip>
-              )
-              : (
-                <Chip size='sm' variant='flat' color='success' className='bg-success-100/50'>
-                  {plugin.version}
-                </Chip>
-              )
-          )}
-      </div>
-
-      {/* 插件类型 */}
-      <div className='hidden sm:block py-3 md:py-4 px-2 md:px-4 text-sm text-center col-span-1'>
-        <div
-          className='w-20 mx-auto h-6 flex items-center justify-center rounded-full border text-xs'
-          style={{
-            backgroundColor: typeConfig.color === 'primary'
-              ? 'rgba(59, 130, 246, 0.1)'
-              : typeConfig.color === 'danger'
-                ? 'rgba(239, 68, 68, 0.1)'
-                : typeConfig.color === 'warning'
-                  ? 'rgba(234, 179, 8, 0.1)'
-                  : 'rgba(161, 161, 170, 0.1)',
-            color: typeConfig.color === 'primary'
-              ? 'rgb(37, 99, 235)'
-              : typeConfig.color === 'danger'
-                ? 'rgb(220, 38, 38)'
-                : typeConfig.color === 'warning'
-                  ? 'rgb(202, 138, 4)'
-                  : 'rgb(113, 113, 122)',
-            borderColor: typeConfig.color === 'primary'
-              ? 'rgba(59, 130, 246, 0.2)'
-              : typeConfig.color === 'danger'
-                ? 'rgba(239, 68, 68, 0.2)'
-                : typeConfig.color === 'warning'
-                  ? 'rgba(234, 179, 8, 0.2)'
-                  : 'rgba(161, 161, 170, 0.2)',
-          }}
-        >
-          {typeConfig.text}
-        </div>
-      </div>
-
-      {/* 状态 */}
-      <div className='py-3 md:py-4 px-0 md:px-4 text-sm text-center col-span-3 sm:col-span-1 pl-1'>
-        <div className='flex justify-end sm:justify-center items-center pr-0 sm:pr-0'>
-          {plugin.type === 'app'
-            ? (
-              <div
-                className='inline-flex items-center justify-center w-20 h-6 text-xs rounded-md border'
-                style={{
-                  backgroundColor: 'rgba(107, 114, 128, 0.08)', /** 灰色背景 */
-                  color: 'rgb(75, 85, 99)', /** 灰色文字 */
-                  borderColor: 'rgba(107, 114, 128, 0.2)', /** 灰色边框 */
-                }}
-              >
-                <TbCircleDashed className='text-xs mr-1' />
-              </div>
-            )
-            : (
-              <div
-                className='inline-flex items-center justify-center w-20 h-6 text-xs rounded-md border'
-                style={{
-                  backgroundColor: statusConfig.color === 'success'
-                    ? 'rgba(34, 197, 94, 0.08)'
-                    : statusConfig.color === 'warning'
-                      ? 'rgba(234, 179, 8, 0.08)'
-                      : 'rgba(161, 161, 170, 0.08)',
-                  color: statusConfig.color === 'success'
-                    ? 'rgb(22, 163, 74)'
-                    : statusConfig.color === 'warning'
-                      ? 'rgb(202, 138, 4)'
-                      : 'rgb(113, 113, 122)',
-                  borderColor: statusConfig.color === 'success'
-                    ? 'rgba(34, 197, 94, 0.2)'
-                    : statusConfig.color === 'warning'
-                      ? 'rgba(234, 179, 8, 0.2)'
-                      : 'rgba(161, 161, 170, 0.2)',
-                }}
-              >
-                {statusConfig.icon}
-                <span className='font-normal'>{statusConfig.text}</span>
-              </div>
-            )}
-        </div>
-      </div>
-
-      {/* 操作 */}
-      <div className='py-3 md:py-4 pl-0 pr-0 sm:px-2 md:px-4 text-sm text-center col-span-2 sm:col-span-1' onClick={stopPropagation}>
-        <div className='flex justify-end sm:justify-center pr-2 sm:pr-0'>
-          <Dropdown>
-            <DropdownTrigger>
-              <Button
-                isIconOnly
-                size='sm'
-                variant='light'
-                color='primary'
-                className='opacity-70 hover:opacity-100 transition-opacity'
-              >
-                <IoSettingsOutline className='text-lg' />
-              </Button>
-            </DropdownTrigger>
-            <DropdownMenu aria-label='插件操作' onAction={handleDropdownAction}>
-              {/* 配置选项 - 只有git类型才显示 */}
-              {plugin.type === 'git'
-                ? (
-                  <DropdownItem key='settings' description='调整插件配置'>
-                    <div className='flex items-center gap-2'>
-                      <IoSettingsOutline />
-                      <span>插件配置</span>
-                    </div>
-                  </DropdownItem>
-                )
-                : null}
-
-              {/* 更新按钮 - 对app类型不显示且必须是可更新状态 */}
-              {plugin.type !== 'app' && plugin.version !== plugin.latestHash
-                ? (
-                  <DropdownItem key='update' description='更新到最新版本' className='text-success'>
-                    <div className='flex items-center gap-2'>
-                      <TbArrowUp />
-                      <span>更新插件</span>
-                    </div>
-                  </DropdownItem>
-                )
-                : null}
-
-              {/* Git类型特有的强制更新按钮 */}
-              {plugin.type === 'git'
-                ? (
-                  <DropdownItem key='forceUpdate' description='强制拉取最新代码' className='text-warning'>
-                    <div className='flex items-center gap-2'>
-                      <TbArrowsUp />
-                      <span>强制更新</span>
-                    </div>
-                  </DropdownItem>
-                )
-                : null}
-
-              {/* 卸载按钮 - 所有类型都显示 */}
-              <DropdownItem key='uninstall' description='从系统移除此插件' className='text-danger'>
-                <div className='flex items-center gap-2'>
-                  <TbTrash />
-                  <span>卸载插件</span>
-                </div>
-              </DropdownItem>
-            </DropdownMenu>
-          </Dropdown>
-        </div>
-      </div>
-    </div>
-  )
-})
-
-PluginRow.displayName = 'PluginRow'
+import { fetchPluginList } from '@/mock/pluginManageData'
+import { FilterCards, TableContent, type PluginType } from '@/components/plugin/admin'
 
 /**
  * 插件管理页面组件
@@ -385,9 +23,6 @@ export const PluginManagePage = (): ReactElement => {
 
   /** 选中的插件ID列表（使用Set保证唯一性和便于操作） */
   const [selectedKeys, setSelectedKeys] = useState(new Set<string>())
-
-  /** 表格容器引用 */
-  const tableContainerRef = useRef<HTMLDivElement>(null)
 
   /**
    * 获取插件列表的请求
@@ -520,51 +155,12 @@ export const PluginManagePage = (): ReactElement => {
    */
   const renderFilterCards = () => {
     const counts = getPluginCounts()
-
     return (
-      <div className='grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4 mb-8'>
-        <StatCard
-          title='全部'
-          count={counts.all}
-          description='全部插件'
-          icon={renderIcon(24, IoSettingsOutline)}
-          gradient='bg-gradient-to-br from-indigo-400/10 to-indigo-500/20 dark:from-indigo-600/10 dark:to-indigo-700/20'
-          border='border border-indigo-200/30 dark:border-indigo-800/20'
-          iconBg='bg-indigo-400/20 dark:bg-indigo-700/30'
-          textColor='text-indigo-700 dark:text-indigo-300'
-          ringColor='ring-indigo-400 dark:ring-indigo-500'
-          isActive={selectedType === 'all'}
-          onClick={() => handleTypeChange('all')}
-        />
-
-        <StatCard
-          title='Git插件'
-          count={counts.git}
-          description='Git  插件'
-          icon={renderIcon(24, TbBrandGit)}
-          gradient='bg-gradient-to-br from-amber-400/10 to-amber-500/20 dark:from-amber-600/10 dark:to-amber-700/20'
-          border='border border-amber-200/30 dark:border-amber-800/20'
-          iconBg='bg-amber-400/20 dark:bg-amber-700/30'
-          textColor='text-amber-700 dark:text-amber-300'
-          ringColor='ring-amber-400 dark:ring-amber-500'
-          isActive={selectedType === 'git'}
-          onClick={() => handleTypeChange('git')}
-        />
-
-        <StatCard
-          title='App插件'
-          count={counts.app}
-          description={window.innerWidth <= 640 ? 'App插件 (插件名称省略了前缀)' : 'App插件'}
-          icon={renderIcon(24, TbApps)}
-          gradient='bg-gradient-to-br from-teal-400/10 to-teal-500/20 dark:from-teal-600/10 dark:to-teal-700/20'
-          border='border border-teal-200/30 dark:border-teal-800/20'
-          iconBg='bg-teal-400/20 dark:bg-teal-700/30'
-          textColor='text-teal-700 dark:text-teal-300'
-          ringColor='ring-teal-400 dark:ring-teal-500'
-          isActive={selectedType === 'app'}
-          onClick={() => handleTypeChange('app')}
-        />
-      </div>
+      <FilterCards
+        counts={counts}
+        selectedType={selectedType}
+        onTypeChange={handleTypeChange}
+      />
     )
   }
 
@@ -572,159 +168,19 @@ export const PluginManagePage = (): ReactElement => {
    * 渲染表格内容
    */
   const renderTableContent = () => {
-    if (loading) {
-      return (
-        <div className='flex items-center justify-center h-60 bg-default-100/30 dark:bg-default-200/10 backdrop-blur-sm rounded-xl shadow-sm'>
-          <Spinner size='lg' color='primary' />
-        </div>
-      )
-    }
-
-    if (!plugins || plugins.length === 0) {
-      return (
-        <div className='flex flex-col items-center justify-center h-60 bg-default-100/30 dark:bg-default-200/10 backdrop-blur-sm rounded-xl shadow-sm'>
-          <p className='text-default-500 mb-4'>没有找到符合条件的插件</p>
-          <Button
-            color='primary'
-            variant='flat'
-            size='sm'
-            startContent={<TbRefresh />}
-            onPress={fetchPlugins}
-          >
-            刷新列表
-          </Button>
-        </div>
-      )
-    }
-
     return (
-      <LazyPluginLoader
+      <TableContent
         plugins={plugins}
-        initialBatchSize={2}
-        batchSize={3}
-      >
-        {({ processedPlugins, progress }) => (
-          <>
-            {/* 显示加载进度 */}
-            {progress < 100 && (
-              <div className='mb-4'>
-                <Progress
-                  value={progress}
-                  color='primary'
-                  size='sm'
-                  radius='sm'
-                  classNames={{
-                    base: 'max-w-md',
-                    track: 'drop-shadow-sm border border-default-100/50',
-                    indicator: 'bg-gradient-to-r from-primary-500 to-primary-400',
-                  }}
-                  aria-label='加载插件进度'
-                />
-                <div className='text-xs text-default-500 mt-1'>
-                  正在加载插件 ({progress}%)... 总计:{plugins.length}个插件
-                </div>
-              </div>
-            )}
-
-            <Card className='overflow-hidden shadow-sm border-none bg-default-50/70 dark:bg-default-100/10 backdrop-blur-sm'>
-              <div className='min-w-full flex flex-col'>
-                {/* 表头 */}
-                <div className='bg-default-50/50 sticky top-0 z-10 w-full'>
-                  <div className='grid grid-cols-12 w-full border-b border-default-100/70 -ml-[3px]'>
-                    {/* 选择框列 */}
-                    <div className='text-default-500 font-normal text-xs uppercase tracking-wider py-2 md:py-4 px-2 md:px-4 flex items-center justify-center col-span-1'>
-                      <div className='pl-1 sm:pl-0'>
-                        <Checkbox
-                          isSelected={isAllSelected()}
-                          onValueChange={handleSelectAll}
-                          size='sm'
-                          aria-label='全选'
-                          classNames={{
-                            base: 'w-4 h-4',
-                            wrapper: 'rounded-full w-4 h-4 border-1 border-default-300 data-[selected=true]:border-blue-500 data-[selected=true]:bg-blue-500 data-[indeterminate=true]:bg-blue-500 data-[indeterminate=true]:border-blue-500 data-[hover=true]:border-blue-400 data-[hover=true]:bg-blue-400/20 transition-all',
-                            icon: 'text-white text-[10px]',
-                          }}
-                        />
-                      </div>
-                    </div>
-
-                    {/* 插件名称列 */}
-                    <div className='text-default-500 font-normal text-xs uppercase tracking-wider py-3 md:py-4 px-2 md:px-4 col-span-6 sm:col-span-4'>
-                      插件名称
-                    </div>
-
-                    {/* 当前版本列 */}
-                    <div className='hidden sm:block text-default-500 font-normal text-xs uppercase tracking-wider py-3 md:py-4 px-1 sm:px-2 md:px-4 text-center col-span-3 sm:col-span-2'>
-                      当前版本
-                    </div>
-
-                    {/* 最新版本列 */}
-                    <div className='hidden sm:block text-default-500 font-normal text-xs uppercase tracking-wider py-3 md:py-4 px-1 sm:px-2 md:px-4 text-center col-span-3 sm:col-span-2'>
-                      最新版本
-                    </div>
-
-                    {/* 插件类型列 */}
-                    <div className='hidden sm:block text-default-500 font-normal text-xs uppercase tracking-wider py-3 md:py-4 px-2 md:px-4 text-center col-span-1'>
-                      插件类型
-                    </div>
-
-                    {/* 状态列 */}
-                    <div className='text-default-500 font-normal text-xs uppercase tracking-wider py-3 md:py-4 px-0 md:px-4 text-center col-span-3 sm:col-span-1'>
-                      <div className='flex justify-end sm:justify-center pr-8 sm:pr-0'>状态</div>
-                    </div>
-
-                    {/* 操作列 */}
-                    <div className='text-default-500 font-normal text-xs uppercase tracking-wider py-3 md:py-4 pl-0 pr-0 sm:px-2 md:px-4 text-center col-span-2 sm:col-span-1'>
-                      <div className='flex justify-end sm:justify-center pr-4 sm:pr-0'>操作</div>
-                    </div>
-                  </div>
-                </div>
-
-                {/* 虚拟滚动容器 */}
-                <div
-                  ref={tableContainerRef}
-                  className='overflow-auto'
-                  style={{ height: `${containerHeight}px` }}
-                >
-                  {/* 内容容器 */}
-                  <div
-                    style={{
-                      height: `${processedPlugins.length * rowHeight}px`,
-                      width: '100%',
-                      position: 'relative',
-                    }}
-                  >
-                    {/* 渲染插件行 */}
-                    {processedPlugins.map((plugin, index) => {
-                      const isSelected = selectedMap.has(plugin.id)
-
-                      return (
-                        <div
-                          key={plugin.id}
-                          style={{
-                            position: 'absolute',
-                            top: `${index * rowHeight}px`,
-                            left: 0,
-                            width: '100%',
-                            height: `${rowHeight}px`,
-                          }}
-                        >
-                          <PluginRow
-                            plugin={plugin}
-                            isSelected={isSelected}
-                            onSelect={handleSelectPlugin}
-                            openSettings={openSettings}
-                          />
-                        </div>
-                      )
-                    })}
-                  </div>
-                </div>
-              </div>
-            </Card>
-          </>
-        )}
-      </LazyPluginLoader>
+        loading={loading}
+        selectedMap={selectedMap}
+        rowHeight={rowHeight}
+        containerHeight={containerHeight}
+        isAllSelected={isAllSelected()}
+        handleSelectAll={handleSelectAll}
+        handleSelectPlugin={handleSelectPlugin}
+        openSettings={openSettings}
+        fetchPlugins={fetchPlugins}
+      />
     )
   }
